app/requirements.txt
Flask-Restless==0.16.0
Flask-JWT==0.3.2
Flask-Security==1.7.4
Flask-Admin==1.3.0
Flask-SQLAlchemy
Flask-Login<0.3
Flask-Mail
Flask-Limiter
flask_paginate
flask-recaptcha
sqlalchemy
redlock-py
psycopg2
twilio
celery
simplifycommerce-sdk-python
supervisor-stdout
tensorflow
numpy
h5py
gensim
geopy
geocoder

